FEAT : GET 메서드 정리
- 에러 처리 함수 작성
- 메서드 헤더 파일에서 cpp 파일로 이전
- 린트 깨져서 수정 핋요
